Moved waiting until socket ready logic into a separate file
